Move Widget functions it its own file. Implement widget_get_window. Implement window_get/set_application.
